[clean]コンポーネントの整理
