+++ src/app/components/user/signup.tsx
@@ -6,59 +6,90 @@ import { useState } from "react"
 import { useForm, SubmitHandler } from "react-hook-form"
 import { zodResolver } from "@hookform/resolvers/zod"
 import Link from "next/link"
-import Loading from "../loading"
+import Loading from "../../loading"
 import * as z from "zod"
-import type { Database } from "../../../lib/database.types"
+import type { Database } from "../../../../lib/database.types"
 type Schema = z.infer<typeof schema>
 
-const schema =z.object({
+const schema = z.object({
+    name: z.string().min(2, { message: "２文字以上入力する必要があります"}),
     email: z.string().email({ message: "メールアドレスの形式ではありません"}),
-    password: z.string().min(6, { message:"6文字以上入力する必要があります"}),  
+    password: z.string().min(6, { message: "6文字以上入力する必要があります"}),
 })
 
-const Login = () => {
+const Signup = () => {
     const router = useRouter()
     const supabase = createClientComponentClient<Database>()
     const [loading, setLoading] = useState(false)
     const [message, setMessage] = useState("")
 
-    const {
+    const{
         register,
         handleSubmit,
         formState: { errors },
+        reset,
     } = useForm({
-        defaultValues:{ email:"", password:"" },
+        defaultValues: { name: "", email: "", password: ""},
         resolver: zodResolver(schema),
     })
 
-    const onSubmit:SubmitHandler<Schema> = async (data) => {
+    const onSubmit: SubmitHandler<Schema> = async (data) => {
         setLoading(true)
-
         try {
-            const { error } = await supabase.auth.signInWithPassword({
+            const { error: errorSignup } = await supabase.auth.signUp({
                 email: data.email,
                 password: data.password,
+                options: {
+                    emailRedirectTo: `${location.origin}/auth/callback`,
+                },
             })
-            if(error){
-                setMessage("エラーが発生しました" + error.message)
+
+            if(errorSignup){
+                setMessage("エラーが発生しました" + errorSignup.message)
                 return
             }
 
-            router.push("/")
+            const { error: updateError } =await supabase
+            .from('profiles')
+            .update({ name: data.name })
+            .eq('email', data.email)
             
+            if(updateError){
+                setMessage("エラーが発生しました" + updateError?.message)
+                return
+            }
+
+            reset()
+            setMessage(
+                "本登録用のURLを記載したメールを送信しました。メールをご確認の上、メール本文中のURLをクリックして本登録を行ってください。"
+            )
         } catch (error) {
             setMessage("エラーが発生しました" + error)
             return
-        } finally{
+        }finally{
             setLoading(false)
             router.refresh()
         }
+
+  
+    
     }
 
     return(
         <div className="max-w-[400px] mx-auto">
-            <div className="text-center font-bold text-xl mb-10">ログイン</div>
+            <div className="text-center font-bold text-xl mb-10">サインアップ</div>
             <form onSubmit={handleSubmit(onSubmit)}>
+                {/* 名前 */}
+                <div className="mb-3">
+                    <input
+                        type="text"
+                        className="border rounded-md w-full py-2 px-3 focus:outline-none focus:border-sky-500"
+                        placeholder="名前"
+                        id="name"
+                        {...register("name", { required: true })}
+                    />
+                    <div className="my-3 text-center text-sm text-red-500">{errors.name?.message}</div>
+                </div>
                 {/* メールアドレス */}
                 <div className="mb-3">
                     <input
@@ -68,7 +99,7 @@ const Login = () => {
                         id="email"
                         {...register("email", { required: true })}
                     />
-                    <div className="my-3 text-center text-sm text-red-500">{errors.email?.message}</div>    
+                    <div className="my-3 text-center text-sm text-red-500">{errors.email?.message}</div>
                 </div>
                 {/* パスワード */}
                 <div className="mb-5">
@@ -79,35 +110,31 @@ const Login = () => {
                         id="password"
                         {...register("password", { required: true })}
                     />
-                    <div className="my-3 text-center text-sm text-red-500">{errors.password?.message}</div>    
+                    <div className="my-3 text-center text-sm text-red-500">{errors.password?.message}</div>
                 </div>
-                {/* ログインボタン */}
+
+                {/*　サインアップボタン */}
                 <div className="mb-5">
                     {loading ? (
                         <Loading />
                     ) : (
                         <button
-                         type="submit"
-                         className="font-bold bg-sky-500 hover:brightness-95 w-full rounded-full p-2 text-white text-sm"
+                            type="submit"
+                            className="font-bold bg-sky-500 hover:brightness-95 w-full p-2 text-white text-sm"
                         >
-                            ログイン
+                            サインアップ
                         </button>
                     )}
                 </div>
             </form>
             {message && <div className="my-5 text-center text-sm text-red-500">{message}</div>}
-            <div className="text-center text-sm mb-5">
-                    <Link href="/auth/reset-password" className="text-gray-500 font-bold">
-                        パスワードを忘れた方はこちら
-                    </Link>
-            </div>
             <div className="text-center text-sm">
-                <Link href="/auth/signup" className="text-gray-500 font-bold">
-                    アカウントを作成する
+                <Link href="/auth/login" className="text-gray-500 font-bold">
+                    ログインはこちら
                 </Link>
+
             </div>
         </div>
     )
 }
-
-export default Login
+export default Signup
