+++ src/app/settings/email/page.tsx
@@ -1,7 +1,7 @@
 import { createServerComponentClient } from "@supabase/auth-helpers-nextjs";
 import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
-import Email from "@/app/components/email"
+import Email from "@/app/components/user/email"
 import type { Database } from "../../../../lib/database.types";
