+++ src/app/settings/logout/page.tsx
@@ -1,7 +1,7 @@
 import { cookies } from "next/headers";
 import { createServerComponentClient } from "@supabase/auth-helpers-nextjs";
 import { redirect } from "next/navigation";
-import Logout from "@/app/components/logout"
+import Logout from "@/app/components/user/logout"
 import type { Database } from "../../../../lib/database.types";
 
 const LogoutPage = async () => {
