+++ src/app/settings/password/page.tsx
@@ -1,7 +1,7 @@
 import { createServerComponentClient } from "@supabase/auth-helpers-nextjs";
 import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
-import Password from "@/app/components/password"
+import Password from "@/app/components/user/password"
 import type { Database } from "../../../../lib/database.types";
