+++ src/app/settings/profile/page.tsx
@@ -1,7 +1,7 @@
 import { createServerComponentClient } from "@supabase/auth-helpers-nextjs";
 import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
-import Profile from "@/app/components/profile"
+import Profile from "@/app/components/user/profile"
 import type { Database } from "../../../../lib/database.types";
 
 const ProfilePage = async () => {
